// src/common/enum.ts
export enum DB_NAME {
  DB_POSTGRES = 'DB_POSTGRES',
}

export const BearerAuthName = 'JWT';

export enum STATUS_UPLOAD {
  INIT = 'INIT',
  PROCESSING = 'PROCESSING',
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
}

export enum ENUM_STATUS {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  INIT = 'INIT',
}

export enum ENUM_MODEL {
  AUTH = 'AUTH',
  USER = 'USER',
  TEACHER = 'TEACHER',
  CLASS = 'CLASS',
  STUDENT = 'STUDENT',
  ATTENDANCE = 'ATTENDANCE',
}

export const I18N_MODEL = {
  [ENUM_MODEL.AUTH]: 'models.AUTH',
  [ENUM_MODEL.USER]: 'models.USER',
  [ENUM_MODEL.TEACHER]: 'models.TEACHER',
  [ENUM_MODEL.CLASS]: 'models.CLASS',
  [ENUM_MODEL.STUDENT]: 'models.STUDENT',
  [ENUM_MODEL.ATTENDANCE]: 'models.ATTENDANCE',
};

export enum ENUM_VALIDATOR {
  IS_STRING = 'IS_STRING',
  IS_NOT_EMPTY = 'IS_NOT_EMPTY',
  LENGTH = 'LENGTH',
  IS_IN = 'IS_IN',
  IS_ENUM = 'IS_ENUM',
  UNIQUE = 'UNIQUE',
  IS_EMAIL = 'IS_EMAIL',
  IS_PASSWORD = 'IS_PASSWORD',
  IS_NUMBER = 'IS_NUMBER',
  MIN = 'MIN',
  IS_EXIST = 'IS_EXIST',
  IS_NUMBER_STRING = 'IS_NUMBER_STRING',
  WRONG_PASSWORD = 'WRONG_PASSWORD',
  CONFIRM_PASSWORD_NOT_MATCH = 'CONFIRM_PASSWORD_NOT_MATCH',
}

export const I18N_VALIDATOR = {
  [ENUM_VALIDATOR.IS_STRING]: 'validation.isString',
  [ENUM_VALIDATOR.IS_NOT_EMPTY]: 'validation.isNotEmpty',
  [ENUM_VALIDATOR.LENGTH]: 'validation.length',
  [ENUM_VALIDATOR.IS_ENUM]: 'validation.isEnum',
  [ENUM_VALIDATOR.UNIQUE]: 'validation.unique',
  [ENUM_VALIDATOR.IS_EMAIL]: 'validation.isEmail',
  [ENUM_VALIDATOR.IS_PASSWORD]: 'validation.isPassword',
  [ENUM_VALIDATOR.IS_NUMBER]: 'validation.isNumber',
  [ENUM_VALIDATOR.MIN]: 'validation.min',
  [ENUM_VALIDATOR.IS_EXIST]: 'validation.isExist',
  [ENUM_VALIDATOR.IS_NUMBER_STRING]: 'validation.isNumberString',
  [ENUM_VALIDATOR.WRONG_PASSWORD]: 'validation.wrongPasswprd',
  [ENUM_VALIDATOR.CONFIRM_PASSWORD_NOT_MATCH]:
    'validation.confirmPasswordNotMatch',
};

export enum ENUM_MESSAGE {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
  CHANGE_STATUS = 'CHANGE_STATUS',
  UNAUTHORIZED = 'UNAUTHORIZED',
  AUTHORIZED = 'AUTHORIZED',
}

export const I18N_MESSAGE = {
  [ENUM_MESSAGE.CREATE]: 'messages.ACTION.CREATE',
  [ENUM_MESSAGE.UPDATE]: 'messages.ACTION.UPDATE',
  [ENUM_MESSAGE.DELETE]: 'messages.ACTION.DELETE',
  [ENUM_MESSAGE.CHANGE_STATUS]: 'messages.ACTION.CHANGE_STATUS',
  [ENUM_MESSAGE.UNAUTHORIZED]: 'messages.AUTH.UNAUTHORIZED',
  [ENUM_MESSAGE.AUTHORIZED]: 'messages.AUTH.AUTHORIZED',
};

export enum ENUM_FIELD {
  CODE = 'CODE',
  AVATAR = 'AVATAR',
  NAME = 'NAME',
  COLOR = 'COLOR',
  STATUS = 'STATUS',
  EMAIL = 'EMAIL',
  PASSWORD = 'PASSWORD',
  PHONE = 'PHONE',
  GENDER = 'GENDER',
  CURRENT_PASSWORD = 'CURRENT_PASSWORD',
  NEW_PASSWORD = 'NEW_PASSWORD',
  CONFIRM_PASSWORD = 'CONFIRM_PASSWORD',
  TYPE = 'TYPE',
  URL = 'URL',
  TEACHER = 'TEACHER',
  CLASSES = 'CLASSES',
}

export const I18N_FIELD = {
  [ENUM_FIELD.CODE]: 'fields.CODE',
  [ENUM_FIELD.AVATAR]: 'fields.AVATAR',
  [ENUM_FIELD.NAME]: 'fields.NAME',
  [ENUM_FIELD.COLOR]: 'fields.COLOR',
  [ENUM_FIELD.STATUS]: 'fields.STATUS',
  [ENUM_FIELD.EMAIL]: 'fields.EMAIL',
  [ENUM_FIELD.PASSWORD]: 'fields.PASSWORD',
  [ENUM_FIELD.PHONE]: 'fields.PHONE',
  [ENUM_FIELD.GENDER]: 'fields.GENDER',
  [ENUM_FIELD.CURRENT_PASSWORD]: 'fields.CURRENT_PASSWORD',
  [ENUM_FIELD.NEW_PASSWORD]: 'fields.NEW_PASSWORD',
  [ENUM_FIELD.CONFIRM_PASSWORD]: 'fields.CONFIRM_PASSWORD',
  [ENUM_FIELD.TYPE]: 'fields.TYPE',
  [ENUM_FIELD.URL]: 'fields.URL',
  [ENUM_FIELD.TEACHER]: 'fields.TEACHER',
  [ENUM_FIELD.CLASSES]: 'fields.CLASSES',
};

export enum ENUM_GENDER {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
  OTHER = 'OTHER',
}